Phase θ: Automerge CRDT substrate for 3+1 multi-user integrations
## CI Fixes (7 failing → 0)
- edu-engine-core: clippy dead_code, assertions_on_constants, unused var
- edu-prover: clippy format_collect, needless_borrows_for_generic_args
- balance-gate: Lean build fallback retry, sorry audit, simulate_battle test
- cross-platform: RISC-V no_std build, no bare-metal linker, aarch64 linker fix

## Phase θ — CRDT Infrastructure
- crates/edu-crdt-bridge: new Rust crate (81 tests: 47 unit + 34 PBT)
  - deck.rs: DeckDocument, DeckEntry, DeckCollection
  - annotation.rs: AnnotationDocument, AnnotationEntry, AnnotationLayer
  - wiki.rs: WikiDocument, WikiPage, WikiNamespace, WikiPageMeta
  - moderation.rs: ModerationQueue, ModerationItem, ModerationAction, ModerationStatus
  - transport.rs: SyncMessage, ChangeSet, ChangeBinary (encode/decode)
  - presence.rs: PresenceInfo, PresenceCursor, DeviceInfo
  - pbt.rs: 34 PBT (convergence, idempotence, commutativity, offline-merge)

## Phase θ — Integration Surfaces
- /decks: cross-device deck sync page + useDeckSync hook
- /replay/[id]: annotation overlay + useAnnotationLayer hook
- /lore/wiki/*: wiki landing + page viewer/editor + useWikiSync hook
- /admin/moderation: moderation queue + useModerationQueue hook
- src/lib/crdt/types.ts: shared TypeScript interfaces

## Deliverables
- ADR-0008: CRDT as substrate (Automerge vs Yjs, deleted surfaces)
- CI: crdt-convergence.yml (convergence simulation, removal tests)
- BENCHMARKS.md: CRDT convergence metrics
- docs/lore-tech-mapping.md: AURALIS Collective Consensus Protocol
- README.md: Server-less multi-user badge

+++ crates/edu-prover/src/merkle.rs
@@ -151,7 +151,7 @@ impl MerkleTree {
         self.root = None; // Invalidate cache
         let n = self.leaves.len();
         if n > 1 {
-            let bits = (usize::BITS - n.leading_zeros()) as u32;
+            let bits = usize::BITS - n.leading_zeros();
             if bits > self.depth {
                 self.depth = bits;
             }
